requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
redis==5.0.1
celery==5.3.4
prometheus-client==0.19.0
sentry-sdk==1.34.0
sqlalchemy==2.0.23
pandas==2.1.3
scikit-learn==1.3.2
python-multipart==0.0.6
pydantic-settings==2.1.0
pydantic==2.5.0
aiohttp==3.9.1
aioredis==2.0.1
python-dotenv==1.0.0
tenacity==8.2.3
flower==2.0.1
structlog==23.2.0
python-json-logger==2.0.7
requests==2.31.0
